style: adjust component layout and add message when there isnt any transaction data
