+++ apps/bytebank/src/app/(pages)/(financial-pages)/_components/extract/_components/delete-modal/index.tsx
